+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Release TBD
   * Implement bulk updates for some product fields.
   * Implement address validation for user-entered addresses.
   * Add Brand and Supplier information for products.
+  * Add tables for regions, countries, states and allow sales restrictions.
 
 ## v1.0.1
 Release 2019-12-24